gRPC_server/requirements.txt
grpcio
grpcio-tools
confluent_kafka
